Asynchronous execution and double precision

+++ src/main/java/org/mmarini/qucomp/compiler/Token.java
@@ -160,7 +160,7 @@ public interface Token {
         /**
          * Returns the value
          */
-        public float value() {
+        public double value() {
             return Float.parseFloat(token());
         }
     }
